Test: GiveawayEvent 클래스 테스트코드 작성
- GiveawayEvent 클래스 내부에 있는 기능들의 테스트코드를 작성.
- 클래스 내부의 인자명이 amount로 되어있던것을 price로 변경.
